pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "app-commands>=0.1.0",
    "discord>=2.3.2",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "typing>=3.10.0.0",
]
